updated settings and user validators, libraries and their corresponding tests
