+++ README.md
@@ -46,7 +46,10 @@ podman run --rm -it localhost/api-server_backend:latest bash -c "npm test"
 
 podman run --rm -it localhost/api-server_backend:latest bash -c "npm --trace-warnings test --detectOpenHandles --verbose " 
 
-podman build -t api-server_backend:latest . && podman run --rm -it   -v /var/home/lbellov/Development/APIServer/API-Server/source:/app:Z   -v node_modules:/app/node_modules:Z   api-server_backend:latest bash
+podman build -t api-server_backend:latest . && podman run --rm -it   -v /home/jesus/Documents/Develop/API-Server/source:/app:Z   -v node_modules:/app/node_modules:Z   api-server_backend:latest bash
+
+### jesus
+podman run --rm -it -v /home/jesus/Documents/Develop/API-Server/source:/app:Z -v node_modules:/app/node_modules:Z api-server_backend:latest bash
 
 ```
