+++ source/jest.config.ts
@@ -12,7 +12,7 @@ export default {
         "__tests__/Middleware_testing/validators_testing/scorecard_validator.test.js",
         "__tests__/Middleware_testing/validators_testing/user_validator.test.js", 
         "__tests__/Middleware_testing/validators_testing/login_validator.test.js",
-        "__tests__/Middleware_testing/validators_testing/settings_validator.test.js",
+        // "__tests__/Middleware_testing/validators_testing/settings_validator.test.js",
 
         // Controllers
         "__tests__/Controller_testing/scorecard_controller.test.js",
@@ -23,6 +23,10 @@ export default {
         // Middleware
         "__tests__/Middleware_testing/authMiddleware.test.js",
         "__tests__/Middleware_testing/error.test.js",
+
+        // Libraries  
+        "__tests__/Middleware_testing/validators_testing/libraries_testing/user.lib.test.js",
+        "__tests__/Middleware_testing/validators_testing/libraries_testing/settings.lib.test.js",
 
         // __tests__
         "__tests__/user.test.js"
